Refactoring ViewPort Part 2
Pulled 2nd half of ShouldAcceptInput into a helper class
